⚙️ refactor(prisma): PrismaModelTypeBuilder createInput and related tests
